+++ src/gui/badge.rs
@@ -2,8 +2,9 @@ use iced::widget::tooltip;
 
 use crate::{
     gui::{
+        common::Message,
         style,
-        widget::{Container, Text, Tooltip},
+        widget::{Button, Container, Text, Tooltip},
     },
     lang::TRANSLATOR,
     scan::ScanChange,
@@ -15,6 +16,7 @@ pub struct Badge {
     left_margin: u16,
     change: Option<ScanChange>,
     tooltip: Option<String>,
+    on_press: Option<Message>,
 }
 
 impl Badge {
@@ -24,6 +26,7 @@ impl Badge {
             left_margin: 0,
             change: None,
             tooltip: None,
+            on_press: None,
         }
     }
 
@@ -72,16 +75,24 @@ impl Badge {
         }
     }
 
+    pub fn on_press(mut self, message: Message) -> Self {
+        self.on_press = Some(message);
+        self
+    }
+
     pub fn view(self) -> Container<'static> {
         Container::new({
             let content = Container::new(Text::new(self.text).size(14))
                 .padding([2, 12, 2, 12])
                 .style(match self.change {
-                    None => style::Container::Badge,
+                    None => match self.on_press.as_ref() {
+                        Some(Message::FilterDuplicates { game: None, .. }) => style::Container::BadgeActivated,
+                        _ => style::Container::Badge,
+                    },
                     Some(change) => style::Container::ChangeBadge(change),
                 });
 
-            match self.tooltip {
+            let content = match self.tooltip {
                 None => content,
                 Some(tooltip) => Container::new(
                     Tooltip::new(content, tooltip, tooltip::Position::Top)
@@ -89,6 +100,11 @@ impl Badge {
                         .gap(5)
                         .style(style::Container::Tooltip),
                 ),
+            };
+
+            match self.on_press {
+                Some(message) => Container::new(Button::new(content).on_press(message).style(style::Button::Badge)),
+                None => Container::new(content),
             }
         })
         .padding([3, 0, 0, self.left_margin])