+++ CHANGELOG.md
@@ -9,6 +9,9 @@
     If a game includes a `.ludusavi.yaml` file,
     then it will be incorporated into the backup scan.
   * Option to sort games by status: new -> different -> same -> unscanned.
+  * GUI: On the backup/restore screen, you can click on the "duplicates" badge next to a game
+    to filter the list down to just the games that conflict with it.
+    You can click the badge again to reset the filter.
   * GUI: On the backup/restore screen, you can use the filter icon to show games
     based on whether they are enabled and whether they have duplicate or ignored saves.
     These filters are reset when you close the program.